backend/requirements.txt
bcrypt==4.0.1
cryptography==38.0.3
decorator==5.1.1
email-validator==1.1.3
fastapi==0.86.0
fastapi-mail==1.2.0
fastjsonschema==2.15.3
httpx==0.23.0
Jinja2==3.0.3
jsonschema==4.4.0
pydantic==1.8.2
PyMySQL==1.0.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==5.0.2
requests==2.26.0
SQLAlchemy-ImageAttach==1.1.0
SQLAlchemy==1.4.43